backend/requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
blinker==1.9.0
branca==0.8.1
bs4==0.0.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
curl_cffi==0.11.4
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
finnhub-python==2.4.24
Flask==3.1.1
flask-cors==6.0.1
folium==0.20.0
frozendict==2.4.6
future==1.0.0
geopandas==1.1.1
h5py==3.14.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
multitasking==0.0.11
narwhals==1.45.0
numpy==2.2.6
packaging==25.0
pandas==2.3.0
peewee==3.18.1
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
protobuf==6.31.1
pvlib==0.13.0
pycparser==2.22
PyJWT==2.10.1
pymongo==4.13.2
PyMySQL==1.1.2
pyogrio==0.11.0
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
reportlab==4.4.3
requests==2.32.4
scikit-learn==1.7.1
scipy==1.15.3
shapely==2.1.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.14.0
tzdata==2025.2
upstox-python-sdk==2.17.0
urllib3==2.5.0
uuid==1.30
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
xyzservices==2025.4.0
yfinance==0.2.64
zipp==3.23.0
